# openapiv2/textgeneration_gateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "textgeneration_gateway.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/textgeneration/multi-upload": {
      "post": {
        "summary": "Upload multiple files and get back a stream of responses",
        "operationId": "TextGeneration_MultiFileUpload",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/textgenerationTextGenResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of textgenerationTextGenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/textgenerationFileUpload"
            }
          }
        ],
        "tags": [
          "TextGeneration"
        ]
      }
    },
    "/v1/textgeneration/single-upload": {
      "post": {
        "summary": "Upload a text file with input prompt and receive a response",
        "operationId": "TextGeneration_SingleFileUpload",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/textgenerationTextGenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/textgenerationFileUpload"
            }
          }
        ],
        "tags": [
          "TextGeneration"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "textgenerationFileUpload": {
      "type": "object",
      "properties": {
        "filename": {
          "type": "string"
        },
        "filedata": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "request message for uploading a file"
    },
...
    "textgenerationTextGenResponse": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        }
      },
      "title": "response message containing output generated by the model"
    }
  }
}